Add administration host prototype
In production, instead of running playbooks directly from technicians
devices, there is an administration host with a dedicated `ansible` user
which owns the playbooks and inventory repositories in their official,
current state, located in `~ansible/playbooks` and `~ansible/inventory`,
respectively.

This is considered less prone to error, e.g. not using a potentially
broken local state which is inaccessible to other team members. It is
also considered more secure, e.g. not requiring a local copy of any
confidential resources in the inventory (though a future version might
rely on a more sophisticated solution than a plain repository).