Add chromium-builder role and chromium-gitlab-runner-0.test host
Introduce the role `chromium-builder` together with a test host and a
playbook. Note that the former is currently not compatible with Debian 10.

The newly introduced playbook prepares a host to become an ABB 2.0
builder, as well as a GitLab runner.

Refer to https://gitlab.com/eyeo/devops/ansible-role-gitlab#synopsis for
details.

This change references https://jira.eyeo.com/browse/OPS-33