Use offical Debian boxes with Vagrant exclusively
Avoid using the "debian/contrib-{buster64,stretch64}" boxes in favor of
the offical "debian/{buster64,stretch64}" versions. The former are not
available for the `libvirt` provider and have been used only to support
the Vagrant shared folder feature on the `virtualbox` provider.

Because we are now using `rsync` for those (and only where necessary),
we can now use the offical boxes that are available for both `libvirt`
and `virtualbox`.