Use offical CentOS boxes with Vagrant exclusively
Avoid using the "generic/centos*" boxes in favor of the offical
"centos/*" versions. The former have been introduced only to support
the Vagrant shared folder feature on the `virtualbox` provider.

Because we are now using `rsync` for those (and only where necessary),
we can now use the offical boxes. Also there is an issue with the
"generic/centos*" ones that prevents proper IP configuration and thus
Vagrant from connecting.