Amend custom cron.d snippet PATH to behave more like generic logrotate
Normally, when logrotate runs, it's started from cron.daily, which in turn
is started from /etc/crontab, which has a default PATH variable set to
include sbin and local binary directory variants.

However, when something is started from cron.d, it seems to have a default
PATH variable unset in configuration, rather it's taken from a cron(8)
compile-time default which excludes those variants. This is the topic of
https://bugs.debian.org/617898

And as it happens, here we need invoke-rc.d to run from /usr/sbin, which
would fail with:

  logrotate_script: 2: logrotate_script: invoke-rc.d: not found

Try defining the default here so it percolates into logrotate.

The legacy implementation of the logrotate-at-midnight didn't have this
bug because it blithely replaced the entire /etc/crontab file in
modules/adblockplus/templates/crontab.erb, which is bad in its own way
(hardcodes even more, affects unrelated logs, replaces a conffile which
can create more issues at package upgrade time, etc)