Implement a basic automated secure backup mechanism
We add a number of building blocks to the backup process: SSH key creation,
SSH authorized keys creation and cron job creation.
Each of these has merged inventory variables that try to be generic, in order
to be reusable outside of the backup logic.
Each of them also has a standalone invocation playbook and a small test host.

The introduction of the SSH key creation tasks required the Ansible
openssh_keypair module, which in turn required raising the minimum Ansible
version overall.

Backup client logic is implemented as a custom_backups dictionary whose elements
include references to both the server and the client variables necessary to make
backup cron jobs work. The dictionary key names are used elsewhere in labels
and filenames. The tasks get delegated to the right place as they go along
based on the context. The dictionary distinguishes between local (client) and
remote (server) users so they don't have to match.

The backup tasks also help integrate SSH between clients and servers by seeding
server public keys into client authorized keys files. During this process,
a temporary file is made in the home directory, to reduce the security risks
from using a shared directory such as /tmp.

The backup clients are allowed to use multiple methods to transfer files,
including scp and rsync. The latter jobs also have log file support.
The backup systems are also tagged to be provisioned with the postfix agents
so that they by default allow mail retrieval of failed cron job logs.

On the server accounts where backups are written, we implement a command
restriction through the force-command modifier of authorized_keys entries,
using a small shell wrapper that is constructed by our templates, allowing only
the commands that match the backup use cases defined and banning everything
else. The wrappers are deployed per client user receiving backups in order
to simplify the output and hopefully reduce the possibility of
mistakes/exploits. The wrappers contain the seemingly complex rsync server flags
that match real-world behaviour - should that behaviour change in the future,
the wrappers would need to be amended.

The example config tries to present all the backup variations based on
conventional directories. It tries to be as DRY as possible, share as many
variables as possible between backup servers and clients using the Ansible
inventory. What would typically be group vars in the production inventory
is instead add custom setup to the Vagrantfile to avoid disturbing inventory
precedence. For this purpose, we also abstract out the global variable setting
mechanism of our Vagrantfile into a helper function, in order to fix ruby
linting issues.

The backup destination directories have to be created as root in case the user
running the backup isn't able to do it (e.g. if they are not part of the group
that is defined in configuration).

We try to explicitly set permissions and ownership/group of various files and
directories involved in setting up backups. This includes the public keys
generated, which have to have owner/group/mode adjusted that way in
Ansible <2.9, otherwise Ansible won't change permissions on them to match
their private key permissions (an issue is already filed upstream).

Backup server default SSH key can be in group_vars in production, but in
the test examples they are in host_vars because we don't want to hardcode
these defaults in a way that production inventory group_vars can't override it.

This implementation follows eyeo/devops/ansible-playbooks.git conventions,
including copying module parameters explicitly, using dictionaries instead
of lists, using dict keys only as fallback to explicit parameters, using
merge_variables with the possibility to deviate suffix, using tasks/paths.yml
custom_path__to_merge instead creating dirs by ourselves, detecting a key
parameter within each dictionary element as a prerequisite to execution
(a kind of a state=ignore), etc.

Inside a few templates that we use, we had to escape the dot, but not just with
a single backslash but with two, otherwise it wouldn't be processed correctly
(Python sucks?).

Cron job tasks don't set the job parameter to mandatory because this would
actually break the module behaviour in the state=absent use case.

When generating output files for cron and authorized_keys entries, we try to
leave some comments to clarify origin in order to help debugging later.

The test VMs need to be able to talk to one another, so we install some
supplementary packages to be able to have ad hoc hostname resolution.
In case of raw vagrant CI, this implies that the client may fail, so
mark it as such for now.

Initially, we also added optional ACL support in order to try to allow backing
up data using an otherwise unprivileged account. However, this seems to have
been a can of worms, so we backed that part out for now.

Ideas and review contributed by Josip Rodin.