Initial attempt for a playbook to install the Eramba ISMS
We tried to honor the Eramba installation documentation from
https://docs.google.com/document/d/1_w38Jb-qiWoDAkOSSgHlrr__bhOHrvoZWu9yinvAEjA/
using our existing task snippets and roles.

The eramba enterprise version is not publicly available, so the tarball needs to be
manually downloaded using the license key and temporarily placed in the inventory.

We have to use two inclusions of tasks/paths.yml due to ordering issues, i.e. some
files can only be created after the source tarball has been extracted and the
directory renamed.

Generic eramba settings have been placed in the group_vars/eramba-servers.yml file
for easy reuse in production inventory.

Our bootstrap plays assume that as long as the directory is empty, it's safe to
import the database (ideally we check that separately, but it's a corner case).