Add a playbook to provision glusterfs clusters
This will add two hosts for Vagrant which need to be brought up simultaneously.

A new variable is introduced for Vagrant, which allows us to change the
Ansible provisioner limit to a group instead of the single host. This behaviour
is required by the gluster module to set up probes and won't work if run
on each host individually.

We rely on the official gluster-ansible-infra Ansible role to set up our
disks and LVM VGs/LVs for use as bricks and eventually volumes.

Avahi packages are added for mDNS, so our Vagrant machines find each other.