Introduce Jirafeau server
This playbook provision a LAMP server and installs the Jirafeau software
with a basic, overridable configuration as well as cron-jobs to remove
expired and asynchronous files.

Known issues:
At the moment the only way to override default dictionaries is to supply all
kay-value pairs, otherwise the unchanged pairs will get lost. We need to figure
out a way to merge the custom dictionary with the default one.