Split product logs for telemetry-pingv1
The goal is making telemetry pingv1 support more products, not just UC. This change splits the SecAuditLog log files by product and updates logrotate to use a date suffix for rotated files.
To simplify inventory files, the templates for the apache vhosts were moved to group_vars and the config files are generated based on a few other vars.