Changed with_dict and with_items to loop according to Ansible 2.5 porting guide
Small style change in apk-signing-host.yml to improve readability

Followed a piece of readability advice by @chacman as original was
cryptic at a glance as a dict inside a jinja variable; and some filter
order improved as well

With the inclusion of loop in favor of with_dict, when the dict index is
a jinja variable, it isn't expanded anymore.
The problem is easily noticeable as the resulting yaml is invalid (in
most of the cases) but can lead to a faulty valued item.

Corrections were made in:
- group_vars/icinga-web-servers.yml
Where it was creating a faulty postgres db name and user. Fixed by
creating a specific variable for these values

- group_vars/ansible-admin-hosts.yml
Where variable was missing. This was fixed by adding the name: "{{
custom_ansible_user.group }}" as suggested by @leunma