Add plays and example for postfix/dkim role
We added a vagrant machine to make the utilization
of postfix/dkim even more clear and not mix it with
crumbs. Because there are dependencies between opendkim
and postfix we created a playbook that can provision
it as a whole.

smtpd_milters and non_smtpd_milters are required for
postfix to forward the outgoing mail to opendkim.
Unfortunately opendkim uses another syntax for the
socket to connect.
We use milter_default_action = accept to prevent
postfix from breaking, in case the opendkim service
croaks.