Introduce geoipupdate tasks
The MaxMind IP database provider actually has a handy program available in
Debian contrib, let's support that for the basic use cases such as telemetry
log parsing

This also makes sure geoipupdate package section is available, because to
actually get the geoipupdate package, we need the base distro's extra repo -
for Debian it's contrib, for Ubuntu it should be multiverse. For anything
else unaccounted for, we leave it presumably broken.