Fix icinga https_response_time check
We bring the changes made in roles/monitoring-plugins (b4ff4093)
here by adding monitoring_plugins_response_time_ssl to the service
configuration for the https check.
There is no necessary adjustment to the http check because the
value for monitoring_plugins_response_time_ssl defaults to false
in the template file.