Integrate sanity check logic for ssh-keys
We did this for tasks/paths.yml as well in !192.
It checks for bad parameters that need to be excepted
when the task is invoked and should fix #74.

Moved this to it's own block to utilize the itemized
variables for all tasks within tasks/ssh-keys.yml

Utilized 'dest' instead of 'path' to be consistent
with task/paths.yml