Use pre-compressed files instead of on-the-fly compression
CPU usage during load was high on filter servers during testing, while nothing substantial happened to memory or disks. Concluded that the server wasting a lot of time for on on-the-fly compression using mod_deflate.
Using pre-compressed files instead of on-the-fly compression using mod_deflate should reduce the load to what we would expect from the new filter servers.