Fixup key-option seeding for authorized_keys
Key options will be dropped, if they are passed in-line
with the key. As this is not intentional behaviour it
needs to be fixed.
Currently the key_options 'default' will set the value
to '', which is not recognized as 'empty' by the module,
so the logic to parse the string is not applied.