Make libvirt-host-0.test CI ready
Volume group creation and a loop device are required
to make this possible. The preparation for this
is done with provision-custom-shells.yml.

Disk space, CPU and memory of the libvirt machines were decreased to
fit within the vagrant machine.

CPU and RAM for libvirt-host-0.test were increased
to speed up the process of vmdebootstrap.

To speed up the process even more, optimizations in terms
of hardware resource allocations were provided by J.Rodin.
This also allows testing on smaller machines, too.

As the nested virtualization aspect is not reliable enough yet,
the CI job is allowed to fail.