Extra versioning requirements for abp-filters-graph serving
This tries to honor an extra specification of:

addonName in ('adblockplus', 'adblockpluschrome', 'adblockplusfirefox', 'adblockplusopera', 'adblockplusedge')
and
addonVersion >= 3.11.4

As requested by Parinitha Hirehal, Uwe Bernitt &co. in HD-3836/OPS-1579

The regular expression is slightly more constrained than that to make it
simple enough while entertaining the most plausible of the possibilities
(we don't expect 3.11.10 to happen, or 3.20+ or 10+ any time soon, cf.
https://eyeo.gitlab.io/adblockplus/abpui/releases/)