openssh_keypair module changed
New openssh_keypair (community.crypto.openssh_keypair) module used to
avoid key overwriting when unneeded by using the "regenerate" parameter.

The default will be 'never' but it could be changed based on variable
name for the task.

The module is available starting ansible 2.10.0, and was tested with
ansible 2.10.17.

The module will always generate a new key if the destination file does
not exist.
If set to never (the default in our configuration) the module will fail
if the key cannot be read or the passphrase isn't matching, and
will never regenerate an existing key.