Adding SSLStaplingCache to global apache2 configuration
To improve performance and privacy for the clients, it has been requested
that the DevOps team enables OCSP Stapling on all apache2 web servers.
This MR adds a variable, to be merged by the ansible plugin, which creates
a file configuration on /conf/local-ssl-stapling-cache.conf that
has SSLStaplingCache, responsible to store cache used by OCSP responses

The other variable added, is used to enable stapling on the domain (virtualhost).
The second file modified is to enable OCSP stapling on the telemetry-ping1-servers,
the virtualhost receives the variable ocsp_stapling_apache_config