Adds support to systemd drop-in feature
Our team wants to know when a systemd service is restarted, therefore a solution
for this is to take advantage of systemd's drop-in directory, where we asdcan add
a configuration, ExecStartPost, that can run a command

The changes on file tasks/service.yml, adds a new variable to be merged
by ansible plugins, this new variable uses the copy module for the configuration file.
This creates a directory in service file directory with the name of the
service plus service.d. Inside of this directory you can find a file
with a name that ends in .conf. This file then holds the configuration
to be append to the service

On host custom-debian11-0.test, the variable can be configured with a name
and a content, in this case, ExecStartPost. It will create a file in the
drop-in directory with the name of local-notifications and our task will
add .conf to the end. Our task will also add what one is specifying
under content or service

To test it, the vagrant machine needs to be brought up and after
provision is done, it is possible to find a conf
file in /etc/systemd/system/example_enabled_service.service.d