Deploy v3 artifacts from web tarball
Download tarball and extract it for further processing.

Append filter to extraction to fetch only needed artifacts
from tarball.

Sanity check index file and fail in case it reports
any issues/inconsistencies. Move artifacts to other artifacts
location, to finalize content for distribution.

Apply retention policy for diff files on v3 artifacts to
delete diff files that are older than X days, so they disappear
in final_output dir and are not transferred further to the filter servers

Make it configurable per item, as we need to cover the probability
of multiple extractable items in an archive and they perhaps need to
be kept longer.