Adds LUKS keys module to tasks directory
With the need to automatically update our LUKS key encryption password for
the disks, we were able to come across an ansible module that we could help us
The ansible luks_device_module the team can easily work with LUKS keys

Since the playbook for custom-disk-0.test creates test partition, I decided to pick
one of the disks to test this module. The partition size had to be changed because
cryptsetup header's size is a 16MB starting on version 2.1.0

A group_vars file for luks-keys have been created so it can hold all the variables
that are needed in every luks-keys host, such as the package "cryptsetup"

The host custom-disk-0.test will add a new password and remove the old one, will also
open and close the disk using the passwords.