Install and configure prerequisites for having syntax checks on ansible inventory files.
Basically to cover the same functionality that we have for
the code repository.

This is achieved by installing pre-commit with yamllint checks.
We apply the same ruleset of ansible-playbooks also for the inventory.

Added instructions in the README, as there are some manual steps
to execute for each local clone of the repository.
Necessary because pre-commit installs a git hook in the repository
that is not copied over to the local checkouts.