Introduce two kinds of Vagrant CI jobs, adding a persistent one in addition to shortlived
A new runner was created to improve the time of our pipeline. This new runner is for
master branch only and it will not delete vagrant machines.

It will verify the status of the machine. If the machine is down or is not created
yet, it will run `vagrant up --no-provision` and after that it will
run `vagrant provision`.

The old runner will still be used but it will run only tests for every commit

Due to this change on our pipeline strategy, the runners had their tags changed and
the pipeline files were updated with the new tags accordingly.

While testing the new pipeline, it was noticed that the jobs needed to have
different names depending on the job, so that someone may be able to easily
noticed if they are shortlived or persistent.