Make merge_variable recurse through nested dictionaries
The previous implementation only supports to merge
dictionaries with a single level, but due to a recent
discovery we need to merge dicts with multiple levels.

Specifically, we need this for tasks/openssh.yml, in order
to override specific group or global settings on a per host
basis, but respect other keys that have been set on higher
level, too.

Therefore, create optional variable to activate recursive merging of
dictionaries and let it default to 'False', so it can be
activated on a play when it's necessary.