Add A/B testing rewrite rules for exceptionrules to filter-server
Add rewrite rules for serving A/B testing filter list if certain
parameters are passed into exceptionrules.txt filter list requests.
This is done as PGU wants to set an A/B test on ABB segmenting
users into different groups on the client side. Based on the groups
the allowlist will be requested with the specified url parameters.