Remove abuse custom apache config for EL-china
After a few syncs with FLOPS and DEVOPS it was mentioned these
config is no longer required.
To validate if this is in fact true, we validated checking filterlists
requests finding not more than 300 requests with status 400, a DOS
attack would result in a much higher number so we can conclude the rules
are not needed anymore.
FLOPS task following up the task: [FLOP-168](https://jira.eyeo.com/browse/FLOP-168).