Remove abuse custom apache config for EL-china
In the process of FLOPS owning filter lists rewrite rules we were
reviewing currently existing rules. We encountered a rule for EL china
that was added with the main focus of prevent DOS attacks, after a few
syncs with FLOPS and DEVOPS it was mentioned these config is no longer
required, and instead of migrating the rule into our htaccess it can be
removed.
To validate if this is in fact true, we reviewed filter lists requests
from the past months finding not more than 300 requests with status 400,
a DOS attack would result in a much higher number so we can conclude the
rules are not needed anymore.