Remove filter delivery server custom apache config
To prepare for FLOPS takeover of filter list rewrite rules in
production, we first want to validate that everything is running and
stable on our staging server.
To be able to validate our new .htaccess file and our rewrite rules are
working properly, we need to remove all filter list related rewrite
rules from the custom apache config.
After the custom apache config is removed we will be able to run all the
test suit in filterlist-delivery, which include some regression testing
to ensure rewrite rules are working properly.

After removing these rules and testing we will be potentially ready for
moving the change to pre prod and eventually production.