Test removal of anti-abuse rule on staging
Removes the custom rule affecting easylistchina+easylist.txt from the
Filter List Delivery system staging environment. This allows the FLOPS
team to further test and validate the rewrite rules generated by their
system.

Further context:
When a request to easylistchina+easylist is made, there is an attempt
to match the rules defined inside the <Location> directive, but
afterwards the request is passed through. A possible explanation for
this behavior can be found in mod_rewrite documentation:
https://httpd.apache.org/docs/2.4/mod/mod_rewrite.html#rewriterule

> Although rewrite rules are syntactically permitted in <Location> and
> <Files> sections (including their regular expression counterparts),
> this should never be necessary and is unsupported. A likely feature
> to break in these contexts is relative substitutions.

It is important to note that this behavior ignores the content of the
.htaccess file, and therefore is not compatible with the changes in
configuration that FLOPS & DEVOPS plan to execute in the near future.
These rules will likely need to be removed from production as well.