System added
option for system user added.

The option is in the original module for creating users, and allows to create a lower than 1000 ID user.
Is it useful when creating a user that will run a service/daemon not to be mixed with the typical user's id

The default (and current only option) is to omit otherwise should be defined to true