Basic prometheus test box
this is a prometheus / alertmanager / mimir skel to work with
all is to be provisioned in prometheus-0.test host but separate with the intention to live in different instances / clusters in real life.
most of the configuration is basic default, to be modified to a still basic default but with more content.

Some observations on the current state:

- there is no obvious correlation / order between user/dir/service creation, is kind of stable, but may give surprises (dirs created before owner, lead to ??? permissions, more or less a workaround is in place)
- alertmanager ui install fails randomly in it's own script, sometimes is to be finished by hand and restart the service.

- grafana will be added later
- node exporter will be added later (by default alertmanager and mimir offers enough to play, though will add it in a next iteration)

while I'm advancing with configuration, I'll add some generics here to fit the test env (like I'll set alertmanager to send emails to a local user in localhost + some random alerts to be triggered, etc)

Nothing (but mimir) looks like changes a lot if not in a real cluster, though alertmanager is configured as a cluster without any other box to communicate with (don't know yet the impact of this, but seems no pain, otherwise would be an HA issue when only one host is alive)

some small identified issues so far:

all services logs in daemon.log (though prometheus creates a dir in /var/log)
all users have /bin/bash as shell by default
prometheus host should have the url in the hosts file to loopback, so it checks whithin the loopback and doesn't go beyond that routing (as is checking in the )http://prometheus-reverse-proxy/metrics endpoint)

Next steps:

adding an alert schema for local basic services and check email in localhost (postfix is installed)
added basic monitoring for local services
check for more advanced apache monitoring
check best way to monitor stuff that doesn't come out of the box nor in typical exporters

Start working on firewalling the whole thing as alertmanager doesn't allow authentication (though could be forced via apache http basic, not sure about how to deal with the cluster thing in port 9094)