Utilize omit instead of None for authorized_key selection
None gets actually evaluated as an empty string ("") and therefore never
did what we wanted.
Usually we relied on the default None when the variable wasn't defined,
but that actually led also to an empty string and caused some weirdness.

Hence, utilize omit as it fits the use case better and is less of a pitfall.
The same logic was also applied to tasks/authorized_keys.yml.

Also move the conditionals into separate lines, as they're evaluated with a logical "and"
operation anyways.

Fixes 234a0674