Introduce elastic role to ansible-playbooks
Invent new elastic-agent-0 to keep the configuration
in .gitlab-ci and Vagrantfile clean.

Add environment variable invocation mechanism to pass the
elastic token and fleet url securely to vagrant.

Increase max lines for Method Length in rubocop.
Number of custom use cases in Vagrant increased and therefore
we need to extend this.

S&P wants to proliferate the utilization and monitoring of machines
and pitched the new role.